Bluetooth (kind of) Works
Bluetooth is properly transmitting data over bluetooth from the microcontroller to the python scrips, but unfortunately it only sends for a few seconds before dropping out.

Works through app on my phone, so it's likely because BT is stricter on computers (timing, control flow, etc).

Not sure what I need to get it working. Have gotten it working through PuTTY before, so likely certain settings in the .serial() function need to be set